feat: add house detail fragment
Adds an onClick listener to the individual house rows, and since from the
API we should have all house data in the initial fetch, we pass this data
directly as arguments to the detail fragment. That is, there's no need
to go to the repo (unless we want to add some update logic later).

This commit also adds a nice transition to the detail fragment and
enables the back arrow in the activity toolbar when the detail fragment
is shown.